routing and cruds

+++ base/views.py
@@ -1,14 +1,68 @@
-from multiprocessing import context
+from email import message
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
+from django.contrib import messages
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Message, Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+
 rooms = [
     {'id': 1, 'name': 'Lets learn python'},
     {'id': 2, 'name': 'Design with me'},
     {'id': 3, 'name': 'Frontend developers'},
 ]
+
+
+def loginPage(request):  # cannot be login as django has prebuilt login method we will use afterwards
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User doesnot exists')
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Invalid credentials")
+
+    context = {'page': page}
+
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Error occured during registration')
+
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request):
@@ -17,20 +71,43 @@ def home(request):
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
-        Q(description__icontains=q) 
-        )
+        Q(description__icontains=q)
+    )
     room_count = rooms.count()
     topics = Topic.objects.all()
-    context = {"rooms": rooms, "topics": topics, "room_count":room_count}
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    context = {"rooms": rooms, "topics": topics,
+               "room_count": room_count, "room_messages": room_messages}
+
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    # query to child model message is model name _set is give us message to this room
+    room_messages = room.message_set.all()
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room': room, 'room_messages': room_messages,
+               'participants': participants}
     return render(request, 'base/room.html', context)
 
+def userProfile(request, id):
+    user = User.objects.get(id=id)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user': user, 'rooms':rooms, 'room_messages':room_messages, 'topics':topics}
+    return render(request, 'base/profile.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,9 +120,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here !!")
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if(form.is_valid):
@@ -56,10 +137,25 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here !!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     context = {'obj': room}
+    return render(request, 'base/delete.html', context)
+
+
+@login_required(login_url='/login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse("You are not allowed here !!")
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    context = {'obj': message}
     return render(request, 'base/delete.html', context)
